+++ Ghost.h
@@ -8,7 +8,7 @@
 
 class Ghost : public Entity, public std::enable_shared_from_this<Ghost> {
 public:
-    Ghost(int x, int y, std::shared_ptr<Pacman> pacman, const std::vector<std::vector<bool>>& wallgrid, int id);
+    Ghost(float x, float y, std::shared_ptr<Pacman> pacman, const std::vector<std::vector<bool>>& wallgrid, int id);
 
     void CalculateNextTurn(float dt);
